requirements.txt
alembic==1.0.5
asn1crypto==0.24.0
backcall==0.1.0
certifi==2019.9.11
cffi==1.12.1
chardet==3.0.4
Click==7.0
croniter==0.3.28
cryptography==2.5
daemonocle==1.0.1
decorator==4.3.0
dnspython==1.16.0
eventlet==0.24.1
Flask==1.0.2
Flask-Login==0.4.1
flask-shell-ipython==0.4.0
Flask-SocketIO==3.1.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
future==0.17.1
greenlet==0.4.15
gunicorn==19.9.0
Jinja2==2.10.3
Mako==1.0.7
MarkupSafe==1.1.1
monotonic==1.5
# numpy==1.16.1
paho-mqtt==1.4.0
# pandas==0.24.0
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
Pygments==2.3.1
python-dateutil==2.7.5
python-dotenv==0.10.0
python-editor==1.0.3
python-engineio==3.9.3
python-socketio==3.0.0
python-telegram-bot==11.1.0
redis==3.0.1
requests==2.22.0
SQLAlchemy==1.3.10
Werkzeug==0.16.0
WTForms==2.2.1
